Remove Accelerator layer and rename core into accelerator
