+++ zigzag/cost_model/cost_model_imc.py
@@ -41,9 +41,8 @@ class CostModelEvaluationForIMC(CostModelEvaluation):
         access_same_data_considered_as_no_access: bool = True,
     ):
         self.is_imc = True
-        self.core = next(iter(accelerator.cores))
-        assert isinstance(self.core.operational_array, ImcArray)
-        self.operational_array: ImcArray = self.core.operational_array
+        assert isinstance(self.accelerator.operational_array, ImcArray)
+        self.operational_array: ImcArray = self.accelerator.operational_array
         super().__init__(
             accelerator=accelerator,
             layer=layer,
