+++ zigzag/stages/exploit_data_locality_stages.py
@@ -9,7 +9,6 @@ from typing import Any
 
 from zigzag.datatypes import MemoryOperand
 from zigzag.hardware.architecture.accelerator import Accelerator
-from zigzag.hardware.architecture.core import Core
 from zigzag.hardware.architecture.memory_hierarchy import MemoryHierarchy
 from zigzag.hardware.architecture.memory_instance import MemoryInstance
 from zigzag.hardware.architecture.memory_level import MemoryLevel, ServedMemDimensions
@@ -36,13 +35,11 @@ class SearchInterLayerDataLocalityStage(Stage):
         super().__init__(list_of_callables, **kwargs)
         self.accelerator = accelerator
         self.workload = workload
-        core_id = accelerator.cores[0].id  # correct only for single-core hardware
-        core = accelerator.get_core(core_id)
         # Remove dummy (non-conv) layers (ReLU, Pooling..) in the layer graph
         self.workload_no_dummy = self.workload.get_copy_no_dummy()
 
         # record of all memory levels
-        self.core_mem_level_list = core.memory_hierarchy.mem_level_list
+        self.core_mem_level_list = accelerator.memory_hierarchy.mem_level_list
         # record of top mem level per layer: dict[layer_idx, mem_level]
         self.mem_update_list: dict[int, dict[MemoryOperand, int]] = {}
         # record of input, output size per layer (unit: bit): dict[layer_idx: [operand: size]]
@@ -228,8 +225,7 @@ class SearchInterLayerDataLocalityStage(Stage):
 
     def check_if_mem_serve_all_oa_dims(self, mem: MemoryLevel, accelerator: Accelerator):
         """! Function to check if mem serve all hardware dimensions"""
-        core = accelerator.cores[0]
-        operational_array = core.operational_array
+        operational_array = accelerator.operational_array
         oa_dim_nb = len(operational_array.dimension_sizes)
         mem_served_oa_dim_nb = len(mem.served_dimensions)
         return mem_served_oa_dim_nb == oa_dim_nb
@@ -303,9 +299,8 @@ class ExploitInterLayerDataLocalityStage(Stage):
     def generate_accelerator_with_removing_unused_memory(self) -> Accelerator:
         # Remove no-use memory level according to update_mem_list and mem_update_weight
         curr_id = self.layer.id
-        core = next(iter(self.accelerator.cores))
-        operational_array = core.operational_array
-        memory_hierarchy = core.memory_hierarchy
+        operational_array = self.accelerator.operational_array
+        memory_hierarchy = self.accelerator.memory_hierarchy
 
         act_layer_op = self.layer.get_act_layer_op()
         weight_layer_op = self.layer.get_weight_layer_op()
@@ -373,20 +368,13 @@ class ExploitInterLayerDataLocalityStage(Stage):
                     served_dimensions=new_served_dimensions,
                 )
 
-        # Create the new core
-        new_core = Core(
-            core_id=core.id,
+        # Create the new accelerator
+        new_name = self.accelerator.name + "-removing-nouse-mem"
+        new_accelerator = Accelerator(
+            core_id=self.accelerator.id,
+            name=new_name,
             operational_array=operational_array,
             memory_hierarchy=new_memory_hierarchy,
-        )
-
-        # Create the new accelerator
-        name = self.accelerator.name
-        new_name = name + "-removing-nouse-mem"
-        new_cores = {new_core}
-        new_accelerator = Accelerator(
-            name=new_name,
-            core_set=new_cores,
         )
 
         logger.info("Update mem architecture for layer %s...", self.layer)
